+++ src/frontend/apps/desk/src/features/teams/components/index.ts
@@ -1,2 +1,3 @@
+export * from './CardCreateTeam';
 export * from './Panel/Panel';
 export * from './TeamInfo';
